improve RunnableSpec
